+++ src/main/java/com/indemnity83/irontank/IronTank.java
@@ -16,40 +16,35 @@ import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 import cpw.mods.fml.common.registry.GameRegistry;
 
-@Mod(modid = Reference.MODID, name = Reference.MODNAME, version = Reference.VERSION)
-public class IronTank
-{
+@Mod(modid = Reference.MODID, name = Reference.MODNAME, version = Reference.VERSION + " build " + Reference.BUILD, dependencies = Reference.DEPENDENCIES)
+public class IronTank {
 
-    @Mod.Instance(Reference.MODID)
-    public static IronTank instance;
+	@Mod.Instance(Reference.MODID)
+	public static IronTank instance;
 
-    @SidedProxy(clientSide = Reference.CLIENT_PROXY_CLASS, serverSide = Reference.SERVER_PROXY_CLASS)
-    public static IProxy proxy;
+	@SidedProxy(clientSide = Reference.CLIENT_PROXY_CLASS, serverSide = Reference.SERVER_PROXY_CLASS)
+	public static IProxy proxy;
 
-    @Mod.EventHandler
-    public void preinit(FMLPreInitializationEvent event)
-    {
-    	ModItems.init();
-    	ModBlocks.init();
-    	LogHelper.info("Pre Initialization Complete!");
-    }
+	@Mod.EventHandler
+	public void preinit(FMLPreInitializationEvent event) {
+		ModItems.init();
+		ModBlocks.init();
+		LogHelper.info("Pre Initialization Complete!");
+	}
 
-    @Mod.EventHandler
-    public void init(FMLInitializationEvent event)
-    {
-    	ModRecipies.init();
-    	LogHelper.info("Initialization Complete!");
-    }
+	@Mod.EventHandler
+	public void init(FMLInitializationEvent event) {
+		ModRecipies.init();
+		LogHelper.info("Initialization Complete!");
+	}
 
-    @Mod.EventHandler
-    public void postinit(FMLPostInitializationEvent event)
-    {
-    	LogHelper.info("Post Initialization Complete!");
-    }
+	@Mod.EventHandler
+	public void postinit(FMLPostInitializationEvent event) {
+		LogHelper.info("Post Initialization Complete!");
+	}
 
-    @Mod.EventHandler
-	public void load(FMLInitializationEvent evt)
-    {
-    	GameRegistry.registerTileEntity(TileIronTank.class, "com.indemnity83.irontank.block.TileIronTank");
-    }
+	@Mod.EventHandler
+	public void load(FMLInitializationEvent evt) {
+		GameRegistry.registerTileEntity(TileIronTank.class, Reference.TILE_IRON_TANK);
+	}
 }