+++ src/main/java/com/indemnity83/irontank/init/ModBlocks.java
@@ -1,27 +1,26 @@
 package com.indemnity83.irontank.init;
 
-import com.indemnity83.irontank.block.BlockIronTank;
-import com.indemnity83.irontank.block.IronTankType;
+import com.indemnity83.irontank.block.BlockExtendedTank;
+import com.indemnity83.irontank.reference.TankType;
 
 import cpw.mods.fml.common.registry.GameRegistry;
 
 public class ModBlocks {
-	
-	public static final BlockIronTank copperTank = new BlockIronTank(IronTankType.COPPER);
-	public static final BlockIronTank silverTank = new BlockIronTank(IronTankType.SILVER);
-	public static final BlockIronTank ironTank = new BlockIronTank(IronTankType.IRON);
-	public static final BlockIronTank goldTank = new BlockIronTank(IronTankType.GOLD);
-	public static final BlockIronTank diamondTank = new BlockIronTank(IronTankType.DIAMOND);
-	public static final BlockIronTank obsidianTank = new BlockIronTank(IronTankType.OBSIDIAN);
-	
-	public static void init()
-	{
-		GameRegistry.registerBlock(ironTank, ironTank.getInternalName());
-		GameRegistry.registerBlock(goldTank, goldTank.getInternalName());
-		GameRegistry.registerBlock(diamondTank, diamondTank.getInternalName());
-		GameRegistry.registerBlock(obsidianTank, obsidianTank.getInternalName());
-		GameRegistry.registerBlock(copperTank, copperTank.getInternalName());
-		GameRegistry.registerBlock(silverTank, silverTank.getInternalName());
+
+	public static final BlockExtendedTank copperTank = new BlockExtendedTank(TankType.COPPER);
+	public static final BlockExtendedTank silverTank = new BlockExtendedTank(TankType.SILVER);
+	public static final BlockExtendedTank ironTank = new BlockExtendedTank(TankType.IRON);
+	public static final BlockExtendedTank goldTank = new BlockExtendedTank(TankType.GOLD);
+	public static final BlockExtendedTank diamondTank = new BlockExtendedTank(TankType.DIAMOND);
+	public static final BlockExtendedTank obsidianTank = new BlockExtendedTank(TankType.OBSIDIAN);
+
+	public static void init() {
+		GameRegistry.registerBlock(ironTank, ironTank.type.name);
+		GameRegistry.registerBlock(goldTank, goldTank.type.name);
+		GameRegistry.registerBlock(diamondTank, diamondTank.type.name);
+		GameRegistry.registerBlock(obsidianTank, obsidianTank.type.name);
+		GameRegistry.registerBlock(copperTank, copperTank.type.name);
+		GameRegistry.registerBlock(silverTank, silverTank.type.name);
 	}
 
 }